+++ src/pages/GetTable/RaceCourse.js
@@ -5,6 +5,8 @@ import { useDispatch, useSelector } from "react-redux";
 import '../../Components/CSS/SubTable.css'
 import { MdDelete } from "react-icons/md";
 import { remove } from "../../redux/postReducer/PostRaceCourse";
+import swal from 'sweetalert';
+
 
 import { Link } from "react-router-dom";
 import Header from "../../Components/Common/Header";
@@ -18,10 +20,28 @@ const News = () => {
   }, [dispatch]);
 
   const handleRemove =  (Id) => {
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this imaginary file!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        swal("Poof! Your imaginary file has been deleted!", {
+          icon: "success",
+        });
+        dispatch(remove(Id));
+      } else {
+        swal("Your imaginary file is safe!");
+      }
+    });
+
 
     
 
- dispatch(remove(Id));
+ 
  
    fetchracecourse();
